Add files via upload
The charge of Mg2+ was taken into account.
Na+ adsorption affinity was adjusted (The adsorption of Na+ is actually higher thant in the experiment but this is necessary to get acceptable loadings of Maleic acid. Furthermore, the amount of adsorbed Mg2+ is ineviatably higher than measured in the experiment.)
Adjustion of the affinities for Maleic acid.